Changed type from long to double in deterministic policies

+++ flink-staging/flink-streaming/flink-streaming-core/src/main/java/org/apache/flink/streaming/api/windowing/policy/DeterministicCountEvictionPolicy.java
@@ -27,7 +27,7 @@ public class DeterministicCountEvictionPolicy<IN> extends CountEvictionPolicy<IN
     }
 
     @Override
-    public long getLowerBorder(long upperBorder) {
+    public double getLowerBorder(double upperBorder) {
         if (upperBorder-maxElements>0){
             return upperBorder-maxElements;
         } else {
